requirements.txt
Flask>=2.0.0
SQLAlchemy>=1.4.0
python-dotenv>=0.19.0
pymysql>=1.0.0
pandas>=1.3.0
openpyxl>=3.0.0
gunicorn>=20.1.0
google-cloud-secret-manager>=2.0.0
google-cloud-logging>=3.0.0
cryptography>=3.4.0
cloud-sql-python-connector>=1.0.0
SQLAlchemy-Utils>=0.37.0
flask-login==0.6.3
